Adding a custom countdown timer.. huge update
I have published a bunch of things here at once.. so basically I made a class for counting seconds and implemented this class inside AlarmActivity with the help of an interface. It is actually basically the same thing I have done for my other activities. I have also implemented the next and previous buttons, which simply increment or decrement the current position value and update the duration values according to that - it has nothing to do with the timer at all. I still have a couple of more things to add to this class.. like an actual alarm at the end of each interval but for the time being I decided not to implement it (it's only like 3 lines of code anyway). I will take some time off now because my exam week is approaching. :) I might write comments or fix small bugs for my other apps but that's going to be about it for some time.